Add check to ensure dconf-tools are installed

+++ patch-mint-x-theme/install.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+command -v dconf >/dev/null 2>&1 || { echo >&2 "I require dconf, part of dconf-tools, but this doesn't exist. Please install dconf-tools: sudo apt-get install dconf-tools"; exit 1; }
 cp -r /usr/share/themes/Mint-X Mint-X-Max
 cd Mint-X-Max
 patch -p1 < ../theme.patch
